Events and the Emitter
Change the way events work. Instead of passing an object of block or
transaction into the event handler method from the event, pass the event
itself. The may be additional data needed in these handlers. In the
handlers, grab the transaction or the block when necessary, such as send
a block through a producer.